Map is now downloading in backgound when app is launched
