+++ MobileAudioGuide/MobileAudioGuide/Resources/infoAboutAuthor.txt
@@ -1,7 +1,3 @@
-Thank you for downloading the Istanbul audio guide app.
-My name is Christopher.
-I am the author of several popular travel websites that help hundreds of thousands of travelers to better plan their trips every year.
-I hope that the Istanbul audio guide will be useful to you on your journey and will help you to get to know one of the most interesting cities in the world much better, significantly saving on guide services.
-If you have comments or suggestions about the application, or if something does not work, or if you want to propose some kind of joint project - send me an e-mail to the email address indicated on this page.
-I try to check my email at least once a day.
-I wish you an exciting trip to Istanbul.
+Christopher guide book author:
+I hope that the Istanbul audio guide will be useful to you on your journey and will help you get to know one of the most interesting cities in the world much better, significantly saving on guide services.
+If you have comments, suggestions or concerns about certain parts of the application, or if you would prefer some kind of joint project, kindly send me an e-mail on the address provided on this page.
